+++ sdk-app/src/design/components/payroll/PayrollEditEmployee/PayrollEditEmployeeStates.tsx
@@ -1,7 +1,7 @@
-import type { Employee } from '@gusto/embedded-api-v-2025-11-15/models/components/employee'
-import type { PayrollEmployeeCompensationsType } from '@gusto/embedded-api-v-2025-11-15/models/components/payrollemployeecompensationstype'
-import type { PayrollFixedCompensationTypesType } from '@gusto/embedded-api-v-2025-11-15/models/components/payrollfixedcompensationtypestype'
-import type { PayScheduleShow } from '@gusto/embedded-api-v-2025-11-15/models/components/payscheduleshow'
+import type { Employee } from '@gusto/embedded-api-v-2026-02-01/models/components/employee'
+import type { PayrollEmployeeCompensationsType } from '@gusto/embedded-api-v-2026-02-01/models/components/payrollemployeecompensationstype'
+import type { PayrollFixedCompensationTypesType } from '@gusto/embedded-api-v-2026-02-01/models/components/payrollfixedcompensationtypestype'
+import type { PayScheduleShow } from '@gusto/embedded-api-v-2026-02-01/models/components/payscheduleshow'
 import {
   PayrollEditEmployeePresentation,
   type WorkweekRange,