+++ src/components/Company/AssignSignatory/CreateSignatory/useCreateSignatory.ts
@@ -1,4 +1,4 @@
-import { type Signatory } from '@gusto/embedded-api-v-2025-11-15/models/components/signatory'
+import { type Signatory } from '@gusto/embedded-api-v-2026-02-01/models/components/signatory'
 import { createCompoundContext } from '@/components/Base'
 import type { RequireAtLeastOne } from '@/types/Helpers'