+++ src/components/Company/StateTaxes/StateTaxesForm/applicableIf.ts
@@ -1,4 +1,4 @@
-import type { TaxRequirement } from '@gusto/embedded-api-v-2025-11-15/models/components/taxrequirement'
+import type { TaxRequirement } from '@gusto/embedded-api-v-2026-02-01/models/components/taxrequirement'
 import { toRhfKey } from './rhfKey'
 
 /**