+++ src/components/Employee/Deductions/shared/useDeductionForm/useDeductionForm.tsx
@@ -6,10 +6,10 @@ import { zodResolver } from '@hookform/resolvers/zod'
 import {
   type Garnishment,
   type GarnishmentType,
-} from '@gusto/embedded-api-v-2025-11-15/models/components/garnishment'
-import { useGarnishmentsCreateMutation } from '@gusto/embedded-api-v-2025-11-15/react-query/garnishmentsCreate'
-import { useGarnishmentsUpdateMutation } from '@gusto/embedded-api-v-2025-11-15/react-query/garnishmentsUpdate'
-import { useGarnishmentsList } from '@gusto/embedded-api-v-2025-11-15/react-query/garnishmentsList'
+} from '@gusto/embedded-api-v-2026-02-01/models/components/garnishment'
+import { useGarnishmentsCreateMutation } from '@gusto/embedded-api-v-2026-02-01/react-query/garnishmentsCreate'
+import { useGarnishmentsUpdateMutation } from '@gusto/embedded-api-v-2026-02-01/react-query/garnishmentsUpdate'
+import { useGarnishmentsList } from '@gusto/embedded-api-v-2026-02-01/react-query/garnishmentsList'
 import {
   createDeductionFormSchema,
   type DeductionFormData,