+++ src/components/Employee/Documents/management/DocumentManager.tsx
@@ -1,6 +1,6 @@
 import { Trans, useTranslation } from 'react-i18next'
-import { useEmployeeFormsGetSuspense } from '@gusto/embedded-api-v-2025-11-15/react-query/employeeFormsGet'
-import { useEmployeeFormsGetPdfSuspense } from '@gusto/embedded-api-v-2025-11-15/react-query/employeeFormsGetPdf'
+import { useEmployeeFormsGetSuspense } from '@gusto/embedded-api-v-2026-02-01/react-query/employeeFormsGet'
+import { useEmployeeFormsGetPdfSuspense } from '@gusto/embedded-api-v-2026-02-01/react-query/employeeFormsGetPdf'
 import { BaseComponent, useBase, type BaseComponentInterface } from '@/components/Base'
 import type { BaseComponentKeys } from '@/components/Base/Base'
 import { ActionsLayout, Flex } from '@/components/Common'