+++ src/test/mocks/apis/contractor_address.ts
@@ -1,8 +1,8 @@
 import type { HttpResponseResolver, PathParams } from 'msw'
 import { http, HttpResponse } from 'msw'
-import type { GetV1ContractorsContractorUuidRequest } from '@gusto/embedded-api-v-2025-11-15/models/operations/getv1contractorscontractoruuid'
-import type { GetV1ContractorsContractorUuidAddressRequest } from '@gusto/embedded-api-v-2025-11-15/models/operations/getv1contractorscontractoruuidaddress'
-import type { ContractorAddressUpdateBody } from '@gusto/embedded-api-v-2025-11-15/models/components/contractoraddressupdatebody'
+import type { GetV1ContractorsContractorUuidRequest } from '@gusto/embedded-api-v-2026-02-01/models/operations/getv1contractorscontractoruuid'
+import type { GetV1ContractorsContractorUuidAddressRequest } from '@gusto/embedded-api-v-2026-02-01/models/operations/getv1contractorscontractoruuidaddress'
+import type { ContractorAddressUpdateBody } from '@gusto/embedded-api-v-2026-02-01/models/components/contractoraddressupdatebody'
 import { getFixture } from '../fixtures/getFixture'
 import { API_BASE_URL } from '@/test/constants'